+++ src/main/java/net/sf/mpxj/primavera/PrimaveraReader.java
@@ -171,11 +171,13 @@ final class PrimaveraReader
       {
          Row row = rows.get(0);
          ProjectProperties properties = m_project.getProjectProperties();
+         properties.setBaselineProjectUniqueID(row.getInteger("sum_base_proj_id"));
          properties.setCreationDate(row.getDate("create_date"));
          properties.setFinishDate(row.getDate("plan_end_date"));
          properties.setGUID(row.getUUID("guid"));
-         properties.setName(row.getString("proj_short_name"));
-         properties.setStartDate(row.getDate("plan_start_date")); // data_date?
+         properties.setProjectID(row.getString("proj_short_name"));
+         properties.setName(row.getString("proj_short_name")); // Temporary, updated later from the WBS
+         properties.setStartDate(row.getDate("plan_start_date"));
          properties.setDefaultTaskType(TASK_TYPE_MAP.get(row.getString("def_duration_type")));
          properties.setStatusDate(row.getDate("last_recalc_date"));
          properties.setFiscalYearStartMonth(row.getInteger("fy_start_month_num"));
@@ -729,13 +731,11 @@ final class PrimaveraReader
       // We set the project name when we read the project properties, but that's just
       // the short name. The full project name lives on the first WBS item. Rather than
       // querying twice, we'll just set it here where we have access to the WBS items.
-      // I haven't changed what's in the project name attribute as that's the value
-      // MPXJ users are used to receiving in that attribute, so we'll use the title
-      // attribute instead.
+      // We'll leave the short name in place if there is no WBS. 
       //
       if (!wbs.isEmpty())
       {
-         projectProperties.setProjectTitle(wbs.get(0).getString("wbs_name"));
+         projectProperties.setName(wbs.get(0).getString("wbs_name"));
       }
 
       //